Feature_PlayerMovement
In this update, a script 'PlayerScript' was created to house commands for the Player.

Node 'PlayerNode' for the Player was created.

PlayerScript and Player Sprite was added to the parent node PlayerNode.

Movement and gravity were tested.

Minor changes to PlayerSprite; removed the collision box outside of the node tree and into PlayerNode, changed the parent node of Player Sprite from CharacterBody2D to Node2D.